# backend/docs/swagger.yaml
basePath: /
definitions:
...
  ai-models-backend_internal_models.ChatMessage:
    properties:
      content:
        type: string
      createdAt:
        type: string
      model:
        type: string
      role:
        type: string
    type: object
  ai-models-backend_internal_models.ChatRequest:
    properties:
      message:
        type: string
      model:
        type: string
      stream:
        type: boolean
    required:
    - message
    type: object
  ai-models-backend_internal_models.ChatResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      message:
        type: string
      model:
        type: string
      usage:
        $ref: '#/definitions/ai-models-backend_internal_models.Usage'
    type: object
...
  ai-models-backend_internal_models.CrudCreateRequest:
    properties:
      category:
        description: 业务分类，默认为general
        type: string
      data:
        type: string
    required:
    - data
    type: object
  ai-models-backend_internal_models.CrudResponse:
    properties:
      category:
        type: string
      createdAt:
        type: string
      data:
        type: string
      id:
        type: string
      updatedAt:
        type: string
    type: object
  ai-models-backend_internal_models.CrudUpdateRequest:
    properties:
      category:
        description: 业务分类
        type: string
      data:
        type: string
    required:
    - data
    type: object
...
  ai-models-backend_internal_models.GenerateRequest:
    properties:
      model:
        type: string
      parameters:
        additionalProperties: {}
        type: object
      prompt:
        type: string
    required:
    - prompt
    type: object
  ai-models-backend_internal_models.GenerateResponse:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      model:
        type: string
      usage:
        $ref: '#/definitions/ai-models-backend_internal_models.Usage'
    type: object
...
  ai-models-backend_internal_models.OpenAIChatCompletionRequest:
    properties:
      maxTokens:
        type: integer
      messages:
        items:
          $ref: '#/definitions/ai-models-backend_internal_models.OpenAIMessage'
        type: array
      model:
        type: string
      platform:
        type: string
      stream:
        type: boolean
      temperature:
        type: number
      topP:
        type: number
    required:
    - messages
    - model
    type: object
  ai-models-backend_internal_models.OpenAIChatCompletionResponse:
    properties:
      choices:
        items:
          $ref: '#/definitions/ai-models-backend_internal_models.OpenAIChoice'
        type: array
      created:
        type: integer
      id:
        type: string
      model:
        type: string
      object:
        type: string
      usage:
        $ref: '#/definitions/ai-models-backend_internal_models.OpenAIUsage'
    type: object
  ai-models-backend_internal_models.OpenAIChoice:
    properties:
      finishReason:
        type: string
      index:
        type: integer
      message:
        $ref: '#/definitions/ai-models-backend_internal_models.OpenAIMessage'
    type: object
  ai-models-backend_internal_models.OpenAIMessage:
    properties:
      content:
        type: string
      role:
        type: string
    required:
    - content
    - role
    type: object
  ai-models-backend_internal_models.OpenAIUsage:
    properties:
      completionTokens:
        type: integer
      promptTokens:
        type: integer
      totalTokens:
        type: integer
    type: object
...
  ai-models-backend_internal_models.Usage:
    properties:
      completionTokens:
        type: integer
      promptTokens:
        type: integer
      totalTokens:
        type: integer
    type: object
...
  ai-models-backend_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: AI Models Backend API documentation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: AI Models API
  version: "1.0"
paths:
...
  /admin/users/{id}/reset-password:
    post:
      description: 管理员重置指定用户的密码，通常用于用户忘记密码或管理员主动重置的场景
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      - description: 新密码
        in: body
        name: request
        required: true
        schema:
          properties:
            newPassword:
              type: string
          type: object
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 重置用户密码
      tags:
      - Admin
  /ai/chat:
    post:
      description: 与AI模型进行对话聊天，支持流式和非流式响应，可以指定使用的AI模型
      parameters:
      - description: 聊天请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.ChatRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.ChatResponse'
              type: object
      summary: AI聊天
      tags:
      - AI
  /ai/chat/history:
    delete:
      description: 清除用户的历史聊天记录，支持按会话ID清除
      parameters:
      - description: 会话ID
        in: query
        name: session_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai-models-backend_pkg_response.Response'
      summary: 清除聊天历史
      tags:
      - AI
    get:
      description: 获取用户的历史聊天记录，支持分页和过滤
      parameters:
      - description: 会话ID
        in: query
        name: session_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ai-models-backend_internal_models.ChatMessage'
                  type: array
              type: object
      summary: 获取聊天历史
      tags:
      - AI
  /ai/generate:
    post:
      description: 基于提示词生成文本内容，支持自定义参数和不同的AI模型
      parameters:
      - description: 生成请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.GenerateRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.GenerateResponse'
              type: object
      summary: 文本生成
      tags:
      - AI
...
  /ai/v1/chat/completions:
    post:
      description: OpenAI兼容聊天接口，支持流式和非流式响应，可以指定使用的AI模型
      parameters:
      - description: 平台
        in: query
        name: platform
        type: string
      - description: OpenAI聊天请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.OpenAIChatCompletionRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ai-models-backend_internal_models.OpenAIChatCompletionResponse'
      summary: OpenAI兼容聊天接口
      tags:
      - AI
  /ai/v1/images/generations:
    post:
      description: 基于提示词生成图片，支持自定义参数和不同的AI模型
      parameters:
      - description: 平台
        in: query
        name: platform
        type: string
      - description: 图片生成请求
        in: body
        name: request
        required: true
        schema:
          properties:
            model:
              type: string
            prompt:
              type: string
          type: object
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      summary: 图片生成
      tags:
      - AI
...
  /crud:
    get:
      description: 分页获取数据记录列表，支持按分类筛选，返回分页信息和记录数据
      operationId: getList
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 分类
        in: query
        name: category
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 获取记录列表
      tags:
      - CRUD
    post:
      description: 创建新的数据记录，支持分类，数据是字符串，前端自行管理决定是否要json parse
      operationId: create
      parameters:
      - description: 创建请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.CrudCreateRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.CrudResponse'
              type: object
      summary: 创建记录
      tags:
      - CRUD
  /crud/{id}:
    delete:
      description: 根据ID永久删除指定的数据记录，操作不可逆，请谨慎使用
      operationId: delete
      parameters:
      - description: 记录ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: 删除记录
      tags:
      - CRUD
    get:
      description: 根据唯一标识符获取指定的数据记录详细信息
      operationId: getByID
      parameters:
      - description: 记录ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.CrudResponse'
              type: object
      summary: 根据ID获取记录
      tags:
      - CRUD
    put:
      description: 根据ID更新现有数据记录的内容，支持更新数据和分类信息
      operationId: update
      parameters:
      - description: 记录ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ai-models-backend_internal_models.CrudUpdateRequest'
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ai-models-backend_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ai-models-backend_internal_models.CrudResponse'
              type: object
      summary: 更新记录
      tags:
      - CRUD
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"